Lower volume levels
I'm not sure if these values will work but I wanted to propose spreading out the volume levels a little more. The loudest setting is fine for very noisy environments but the mid and low settings could be lower in my opinion. The most important change is to the low setting since it just seems a bit too loud for a quiet environment where you want sound but don't want it overly loud.